Fix bug in screenshot DSL parsing
I guess I was expecting `String.prototype.split` to work this way:

```js
> "click .rentals li:first-of-type article.rental .image".split(/\s+/, 2);
[ 'click', 'rentals li:first-of-type article.rental .image' ]
```

But it does not, it returns this:

```js
> "click .rentals li:first-of-type article.rental .image".split(/\s+/, 2);
[ 'click', '.rentals' ]
```

TL;DR I was expecting the limit paramemter to mean that it will only
split the first space, and the rest will be returned unsplit (like how
it is in Ruby). However, it actually just split all the spaces anyway,
the "extra" strings are just dropped from the returned array.