Reuse build from `ember s` in most tests
We currently spend a lot of time running `yarn test`, which needs
to do a full build before running the tests. Since we already have
the ember server running in the background, in most cases we should
be able to just reuse the build a save a lot of time.