Temporarily disable image optimization
The image optimization setting is known to be pretty slow. This
is not usually a problem because the step is not performed when
the images didn't change. Most of the time, at most one or two
images will change per build so that's not a big issue.

With the flood of new images added in part 2 (#75), this step is
now pushing us over the 50 min time limit on Travis. This commit
disable the step for now so it can be run manually outside of CI.

This commit should be reverted once the build is green again.