Make screenshot less flaky
Add a second parameter to the `click` screenshot DSL to denote
"external" links that performs a navigation. This will wait for
the navigation to complete before moving on to the next step and
hopefully alleviate some of the build failures we are seeing with
puppeteer reporting "Cannot take screenshot with 0 width".