More aggressive downsampling in perceptual diff
The screenshots we take are quite large, and the diff algorithm
often picks up details you won't notice unless you are really
looking at the before/after side-by-side with a magnifying glass.

Usually the changes are due to updates in Open Street Map data that
causes the MapBox maps to change slightly (adding or removing a
street, that kind of stuff).