Fix race conditions with git
The `xargs` invocation in the image diffing step runs two tasks at
the same time. When these two tasks try to access the git index
concurrently, git will error, causing the function to think that
the image is new (as it "didn't exist" on the current branch). This
commit fixed the issue by first checking out the current branch to
a temporary location before the parallel `xargs` is inovoked.