Added missing .jumbo class for consistency's sake
In all the other assertions for the other pages we also use the ´.jumbo´ class. While I understand this still works correctly for the tests in the tutorial, I feel like it is unnecessarily confusing for beginners (especially when it's not called out in the text after teh example either).